+++ examples/delReserve.ts
@@ -5,7 +5,7 @@
  */
 import * as COA from '../lib/index';
 
-COA.ReserveService.delReserve({
+COA.services.reserve.delReserve({
     reserve_num: 985,
     theater_code: '118',
     date_jouei: '20170403',