+++ examples/findPerformances.ts
@@ -5,7 +5,7 @@
  */
 import * as COA from '../lib/index';
 
-COA.MasterService.schedule({
+COA.services.master.schedule({
     theater_code: '118',
     begin: '20170411',
     end: '20170411'